python/Document_processing_api/requirement.txt
flask
flask-cors
pypdf
pymilvus
openai
python-dotenv
